+++ src/nespresso/recsys/preprocessing/model.py
@@ -9,17 +9,17 @@ EMBEDDING_LEN = 768
 _MODEL_NAME = "sentence-transformers/all-mpnet-base-v2"
 
 
-def EnsureEmbeddingModel() -> None:
+def EnsureSentenceTransformer() -> None:
     snapshot_download(
         repo_id=_MODEL_NAME,
         local_dir=DIR_EMBEDDING,
     )
 
 
-def LoadEmbeddingModel() -> SentenceTransformer:
-    EnsureEmbeddingModel()
+def GetSentenceTransformer() -> SentenceTransformer:
+    EnsureSentenceTransformer()
 
     return SentenceTransformer(model_name_or_path=str(DIR_EMBEDDING))
 
 
-embedding_model = LoadEmbeddingModel()
+model = GetSentenceTransformer()